+++ res/shaders/selection/selection_vertex.vert
@@ -15,7 +15,7 @@ uniform vec3 u_cam_pos;
 uniform float u_cell_size;
 uniform float u_average_cell_size;
 
-uniform int u_peel_depth;
+uniform float u_peel_depth;
 uniform float u_slice_depth;
 uniform vec3 u_min;
 uniform vec3 u_max;
